fix: VoskRecognizer sampleRate parameter and Retrofit kapt annotations
- Fix VoskRecognizer: pass Float directly to SpeechService constructor
- Add explicit Retrofit kapt dependencies to data module for annotation processing

+++ extensions/src/main/java/com/kurostream/extensions/asr/VoskRecognizer.kt
@@ -57,7 +57,7 @@ class VoskRecognizer @Inject constructor(private val context: Context) {
     fun startListening(sampleRate: Float = 16000f): Flow<AsrResult> = callbackFlow {
         val currentModel = model ?: run { trySend(AsrResult.Error("Model not initialized")); close(); return@callbackFlow }
         val recognizer = Recognizer(currentModel, sampleRate)
-        speechService = SpeechService(recognizer, sampleRate.toInt())
+        speechService = SpeechService(recognizer, sampleRate)
         speechService?.startListening(object : org.vosk.android.RecognitionListener {
             override fun onPartialResult(hypothesis: String?) { hypothesis?.let { _partialResults.value = it } }
             override fun onResult(hypothesis: String?) { hypothesis?.let { _transcription.value += " $it"; trySend(AsrResult.Partial(it)) } }
